requirements.txt
python-dotenv>=1.0.0
openai>=1.12.0
colorama>=0.4.6
PyYAML>=6.0.2 